Remove inherent methods from `coverageinfo::ffi::Counter`
This patch also moves `Regions` to a different module, since that type can stay
put when the FFI bindings move to another crate.
